+++ src/main/java/dev/amot/endshards/tools/BaseToolMaterial.java
@@ -1,23 +1,11 @@
-package dev.amot.endshards;
+package dev.amot.endshards.tools;
 
 import net.minecraft.item.Item;
 import net.minecraft.item.ToolMaterial;
 import net.minecraft.recipe.Ingredient;
 
-public class BaseToolMaterial implements ToolMaterial {
-    private final Item repair_ingredient;
-    private final int durability;
-    private final float mining_speed_multiplier;
-    private final int mining_level;
-    private final int enchantability;
-
-    BaseToolMaterial(Item repair_ingredient, int durability, float mining_speed_multiplier, int mining_level, int enchantability){
-        this.repair_ingredient = repair_ingredient;
-        this.durability = durability;
-        this.mining_speed_multiplier = mining_speed_multiplier;
-        this.mining_level = mining_level;
-        this.enchantability = enchantability;
-    }
+public record BaseToolMaterial(Item repair_ingredient, int durability, float mining_speed_multiplier,
+                               int mining_level, int enchantability) implements ToolMaterial {
 
     @Override
     public Ingredient getRepairIngredient() {
